test(simulation): split service tests by feature
